Deleção de item do document pedido

+++ src/main/java/br/com/kommando/item/data/services/ItemService.java
@@ -1,8 +1,11 @@
 package br.com.kommando.item.data.services;
 
+import br.com.kommando.exception.error.DataFoundException;
 import br.com.kommando.exception.error.DataNotFoundException;
 import br.com.kommando.item.data.models.Item;
 import br.com.kommando.item.repository.ItemRepository;
+import br.com.kommando.pedido.data.models.Pedido;
+import br.com.kommando.pedido.repository.PedidoRepository;
 import br.com.kommando.produto.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,13 +22,25 @@ public class ItemService {
     @Autowired
     private ProdutoRepository produtoRepository;
 
+    @Autowired
+    private PedidoRepository pedidoRepository;
+
     public List<Item> findAll() {
         return repository.findAll();
     }
 
-    public Item save(Item item) {
+    public Item save(String pedidoId, Item item) {
+        Optional<Pedido> pedidoOptional = pedidoRepository.findById(pedidoId);
+        if (pedidoOptional.isEmpty()) {
+            throw new DataNotFoundException("Pedido não encontrado");
+        }
         if (produtoRepository.findById(item.getProdutoId()).isPresent()) {
-            return repository.save(item);
+            Pedido pedido = pedidoOptional.get();
+            item.setPedidoId(pedidoId);
+            Item itemSalvo = repository.save(item);
+            pedido.getItems().add(itemSalvo);
+            pedidoRepository.save(pedido);
+            return itemSalvo;
         } else {
             throw new DataNotFoundException("Produto não encontrado");
         }
@@ -37,7 +52,28 @@ public class ItemService {
 
     public void removeById(String id) {
         try {
-            repository.deleteById(id);
+            Optional<Item> itemOptional = repository.findById(id);
+            if (itemOptional.isPresent()) {
+                Item item = itemOptional.get();
+                Optional<Pedido> pedidoOptional = pedidoRepository.findById(item.getPedidoId());
+                int index = -1;
+                if (pedidoOptional.isPresent()) {
+                    for (int i = 0; i < pedidoOptional.get().getItems().size(); i++) {
+                        if (pedidoOptional.get().getItems().get(i).getId().equals(item.getId())) {
+                            index = i;
+                            repository.deleteById(id);
+                        }
+                    }
+                    if (index != -1) {
+
+                        pedidoOptional.get().getItems().remove(index);
+                    }
+                    pedidoRepository.save(pedidoOptional.get());
+                }
+            } else {
+                throw new DataFoundException("Item não existe!");
+            }
+
         } catch (Exception e) {
             throw new IllegalArgumentException();
         }
